Implement set/get pointer focus
This commit implements ivi_layout_SetPointerFocusOn and
ivi_layout_GetPointerFocusSurfaceId.

Note that ivi_layout_SetPointerFocusOn changes the position
of the pointer, as well as setting the pointer focus, because
otherwise the pointer focus will be immediately changed back to
the surface the pointer was previously inside (or no surface at all).

Signed-off-by: James Thomas <james.thomas@codethink.co.uk>