ivi-layout: Introducing new notification for render order changes.
the surfaces/layers should get IVI_NOTIFICATION_ADD if they are added to a layer, and they should get
IVI_NOTIFICATION_REMOVE if they are removed from a layer/screen.
Pending lists should be compared with order lists to ensure the notifications are correct.
Because a layer/surface should not get Add notification if they are already on the screen/layer.
A layer/surface should also not get Remove notification if they are not on the screen/layer in the first place.

Signed-off-by: Ahmet Emre Ucan <eucan@de.adit-jv.com>