protocol: add interface ivi_application extension (v2)
ivi_application extension creates ivi_surface objects tied
to a given wl_surface with a given id. The given id can be used in a
shell to identify which application is assigned to a wl_surface and
layout the surface wherever the shell wants. ivi_surface objects can
be used to receive status of wl_surface in the scenegraph of the
compositor.

Signed-off-by: Ahmet Emre Ucan <eucan@de.adit-jv.com>