ivi-layout: Fix surface masks not being reset on layer change
This is not the most efficient way of determining when to regenerate
surface masks (If the surface is moved to an identical layer, it'll
regenerate the mask when it didn't need to).

A more efficient way would be to compare the current layer's properties
to the next layer's pending properties, but that is a non-trivial change.