Initialize the localInfo pointer
This missing initialization causes gcoOS_InitLocalDisplayInfo
api to fail with a segfault.

Additionally, if several displays are recognized, gcoOS_InitLocalDisplayInfo()
will be called several times at the moment. As a result of it, segfault had been occurred by the second call.

This patch avoid to call it several times as well.

Signed-off-by: harsha.mm <harsha.mm@de.adit-jv.com>
Signed-off-by: Hiroshi Noda <hnoda@jp.adit-jv.com>
Signed-off-by: Nobuhiko Tanibata <ntanibata@jp.adit-jv.com>