compositor:imx6drm: don't execute the drmModeSetCrtc if the drm mode and buffer are compatible
To avoid flickering on display during compositor startup, imx6 drm/kms implementation
has a optimization in the drmModeSetCrtc api:
if the mode and drm buffer are compatible drm framework will just execute a page flip in the kernel
and not a full mode setting (which will lead to a flicker).
This implementation make the drmModeSetCrtc api very fast in this special scenario.
The problem is that weston will call directly after drmModeSetCrtc drmModePageFlip api
but the page flip is not executed an the api will fail, also there is no notification installed to wait for the page flip
The solution provided by this patch is to check if drm mode a drm buffer are compatible
and skip the drmModeSetCrtc in this special case.

Signed-off-by: Eugen Friedrich <efriedrich@de.adit-jv.com>