ivi-shell: add ivi-layout API for remoting
Add ivi-layout API to start and stop remoting of a surface, a query
whether a surface is remoted, and a listener API for notifying about
per-surface stream status changes.

Add shell-facing ivi-layout API for informing of configure calls.
It seems it's best to add a new entry point to ivi-layout for this,
since ivi_layout_surface_configure() is meant to re-layout and that
might be heavy.

It also looks like ivi-shell is the designed place to check if the
surface size (from client perspective) changed.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>