compositor: add weston_surface_force_output()
This will ensure that a surface stays synchronized to the given output,
regardless of intersection areas with all outputs.

The forced output is also inserted in the surface's output mask. If you
force a surface to be synchronized to a particular output, then
certainly that surface will be visible on that output.

Forcing the sync output is useful in special cases, e.g. per-output
surfaces (background, panel?) should probably be locked to the output
they are meant for. So far this has not been an issue, because we have
not had overlapping output regions, so assignment by coverage always
produces the right result.

Another type of special cases, which actually prompted for this
feature, are special outputs that are not part of the global coordinate
space at all. The Transmitter plugin (not yet upstream) will make use of
this feature.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>