input: add support for external pointer focus management
Add signal for notifying when clients create a new wl_pointer object,
and export weston_pointer_get_pointer_client().

These are used by alternative input implementations that do not go through
the global input layers but maintain pointer focus externally. That is,
implementation that cannot use weston_pointer::focus.

A user of this is the Transmitter plugin (not yet upstream) which
hijacks surfaces and puts them on a remote display. The input for those
surfaces comes from the remote, and therefore also focus management is
in the remote.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.co.uk>