Set number of display buffers to 3
This is done as a workaround to a bug in imx6 ipu.
imx6 ipu can report page flip completion even if the
page flip is not complete. This can happen when
interrupt latency is higher. Enabling triple buffering
gives an additional buffer for weston to renderi, thereby
avoiding flicker on the display.

Signed-off-by: harsha.mm <harsha.mm@de.adit-jv.com>