modified the commit_changes
If the visibility for a surface/layer is set as 0, then the dirty rect is
always set as empty but repaint is scheduled from commit_changes. from weston
point of view frame n+1 is not getting the update from frame n.

To avoid this, added weston_view_damage_below() API for the view(s) to be hidden

Signed-off-by: Maniraj Devadoss <external.mdevadoss@de.adit-jv.com>