configure.ac: Fix wayland-protocols path (ADIT)
The wayland-protocols directory is used during build: Fix the path
to point to sysroot specified in recipe.

Normally PKG_CONFIG_SYSROOT_DIR could be used in configure.ac but that
breaks multilib weston as it would point to multilib sysroot when the
(allarch) wayland-protocols is actually in the machine sysroot.

I updated it for weston 2.0

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>