launcher: Fix seg-fault on weston exit
The previous commit:

f17176bb9ecf ("Run weston as system user instead of root")

causes a segmentation fault when weston exits:

[19:00:43.549] caught signal: 11
[19:00:43.549]   [000000000041e35a]  --  (weston)
[19:00:43.549]   [000000000041e3c7]  --  (weston)
[19:00:43.549]   [00000033eac330e0]  --  (/lib64/libc.so.6)
[19:00:43.549]   [00000033f2a092e5]  wl_event_source_remove  (/usr/lib64/libwayland-server.so.0)
[19:00:43.549]   [00007fa78deb2e56]  --  (/usr/lib64/weston/drm-backend.so)
[19:00:43.549]   [00007fa78deabf48]  --  (/usr/lib64/weston/drm-backend.so)
[19:00:43.549]   [0000000000412509]  weston_compositor_destroy  (weston)
[19:00:43.549]   [0000000000408d21]  --  (weston)
[19:00:43.549]   [00000033eac202c1]  __libc_start_main  (/lib64/libc.so.6)
[19:00:43.549]   [000000000040a42a]  _start  (weston)

Specifically this was caused by de-referencing launcher->vt_source which
was not set because setup_tty() was skipped. Fix this by calling setup_tty()
for root users as before, and then check for launcher->tty >=0 and
launcher->vt_source != NULL before their usage in case of non-root users.

It's not clear why the previous commit disabled setup_tty() for non-root
users, but likely the reason is that non-root users require the
CAP_SYS_TTY_CONFIG capability to configure tty's.

Fixes: f17176bb9ecf ("Run weston as system user instead of root")

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>